+++ src/server/game/AuctionHouse/AuctionHouseMgr.h
@@ -60,11 +60,11 @@ struct AuctionEntry
     uint32 item_guidlow;
     uint32 item_template;
     uint64 owner;
-    uint64 startbid;                                        //maybe useless
-    uint64 bid;
-    uint64 buyout;
+    uint32 startbid;                                        //maybe useless
+    uint32 bid;
+    uint32 buyout;
     time_t expire_time;
-    uint64 bidder;
+    uint32 bidder;
     uint32 deposit;                                         //deposit can be calculated only when creating auction
     AuctionHouseEntry const* auctionHouseEntry;             // in AuctionHouse.dbc
     uint32 factionTemplateId;
